Added Pocoyo Candy Drop
This Scratch project is a game involving Pocoyo, the titular character of a popular children's television show in Spain. Candy falls from the sky. The player must click the candy to eat it before Pocoyo does. (Before it touches Pocoyo.) 
Click the candy to eat it, press the spacebar to pause the game, and click the volume button to mute or unmute the music. Eat 50 pieces of candy without any of them touching Pocoyo, and you win!

P.S. There is an Easter egg involving a sleepy bird from the show.

Enjoy!